requirements_openpilot.txt
# This are the packages installed on the EON
<<<<<<< HEAD
-e git+https://github.com/commaai/le_python.git#egg=Logentries
-e git+https://github.com/commaai/python-overpy.git#egg=overpy
Cython==0.27.3
Flask==1.0.2
#PyGObject==3.28.2 This is installed on the EON, but requires a ton of dependencies to install
=======
-e git+https://github.com/commaai/le_python.git@5eef8f5be5929d33973e1b10e686fa0cdcd6792f#egg=Logentries
-e git+https://github.com/commaai/python-overpy.git@f86529af402d4642e1faeb146671c40284007323#egg=overpy
Cython==0.27.3
Flask==1.0.2
#PyGObject==3.28.2 This is installed on the EON, but requires a ton of dependencies to install
PyJWT==1.4.1
>>>>>>> 7d5332833b11570db288f35657a963ed0d8cad0a
PyYAML==3.12
appdirs==1.4.0
atomicwrites==1.1.5
attrs==16.0.0
bitstring==3.1.5
capnpy==0.4.2
certifi==2016.8.31
cffi==1.11.5
contextlib2==0.5.4
crc16==0.1.1
crcmod==1.7
cryptography==1.4
cycler==0.10.0
decorator==4.0.10
docopt==0.6.2
enum34==1.1.6
evdev==0.6.1
fastcluster==1.1.20
filterpy==1.2.4
hexdump==3.3
ipaddress==1.0.16
json-rpc==1.12.1
libusb1==1.5.0
lmdb==0.92
mpmath==1.0.0
nose==1.3.7
numpy==1.11.1
opencv-python==3.4.0.12
pause==0.1.2
psutil==3.4.2
py==1.4.31
pyOpenSSL==16.0.0
pyasn1-modules==0.0.8
pyasn1==0.1.9
pycapnp==0.6.3
pycparser==2.18
pycrypto==2.6.1
pyflakes==1.6.0
pylint==1.8.3
pyopencl==2016.1
pyparsing==2.1.10
#pypcap==1.1.5 needs extra dependencies and is not used
pyproj==1.9.5.1
pypytools==0.4.3
pyserial==3.1.1
pytest==2.9.2
python-dateutil==2.6.0
pytools==2016.2.1
pytz==2016.10
pyyaml==3.12
pyzmq==15.4.0
raven==5.23.0
recordclass==0.4.1
requests==2.10.0
scipy==0.19.1
service-identity==16.0.0
setproctitle==1.1.10
simplejson==3.8.2
six==1.10.0
smbus-cffi==0.5.1
smbus2==0.2.0
sympy==1.1.1
tqdm==4.23.1
ujson==1.35
v4l2==0.2
websocket_client==0.55.0
